+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   Fixed ClickAwayProvider initialization error using deferred initialization.
+-   Fixed `Popover` placement on children update.
 
 ## [0.28.0][] - 2020-11-17